FruitStore 싱글톤으로 변경 및 notification 테스트
