+++ JuiceMaker/JuiceMaker/Controller/FruitStoreVC.swift
@@ -7,28 +7,28 @@
 
 import UIKit
 
-class FruitStoreVC: UIViewController {
-
-    @IBAction func goJuiceMaker(_ sender: UIButton) {
-        dismiss(animated: true)
-    }
-    
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
+//class FruitStoreVC: UIViewController {
+//
+//    @IBAction func goJuiceMaker(_ sender: UIButton) {
+//        dismiss(animated: true)
+//    }
+//    
+//    
+//    override func viewDidLoad() {
+//        super.viewDidLoad()
+//
+//        // Do any additional setup after loading the view.
+//    }
+//    
+//
+//    /*
+//    // MARK: - Navigation
+//
+//    // In a storyboard-based application, you will often want to do a little preparation before navigation
+//    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+//        // Get the new view controller using segue.destination.
+//        // Pass the selected object to the new view controller.
+//    }
+//    */
+//
+//}
