+++ JuiceMaker/JuiceMaker/Model/JuiceMaker.swift
@@ -5,12 +5,7 @@
 // 
 
 class JuiceMaker {
-    private var fruitStore = FruitStore(stock: 10)
-    
-    func checkStock(fruit: Fruit) -> Int {
-        let stock = fruitStore.checkFruitNumber(fruit: fruit)
-        return stock
-    }
+    private var fruitStore = FruitStore.shared
     
     func takeOrder(order: Menu) throws {
         let recipe = fetchRecipe(menu: order)
